==> 3. SUBRELOBJ CSV Analysis.ipynb <==
import pandas as pd
import random 
import matplotlib.pyplot as plt
from tqdm import tqdm

### Read the CSV file

## Load the file SUBRELOBJ.csv from https://zenodo.org/records/10022727

SUBRELOBJ = pd.read_csv('SUBRELOBJ.csv')

Analyzing the DataFrame

print("Number of rows: {}".format(SUBRELOBJ.shape))
print("Columns : {}".format(SUBRELOBJ.columns.tolist()))

Total unique entities

unique_entities = list(set(SUBRELOBJ['Subject'].tolist() + SUBRELOBJ['Object'].tolist()))
print("Number of unique entities : {}".format(len(unique_entities)))

Let's see a couple of random samples

SUBRELOBJ.sample(5)

### Plotting the distribution of relationships

rel_count = SUBRELOBJ.groupby('Rel').size().reset_index(name = 'count').sort_values('count', ascending = False)
rel_count.head(25)

rel_count.tail(5)

### Querying CSV File

def query_subrelobj(query, rel, k=10):
    
    def get_relevant_entities(query):
        entities = []
        for ent in unique_entities:
            if query.lower() in ent.lower():
                entities.append(ent)
        return entities
    
    def query_df(entities, rel):
        filtered_df = SUBRELOBJ[(SUBRELOBJ['Subject'].isin(entities)) & (SUBRELOBJ['Rel'] == rel)]
        grouped_df = filtered_df.groupby('Object')['Count'].sum().reset_index()
        sorted_df = grouped_df.sort_values(by='Count', ascending=False)
        objects_and_counts = sorted_df.set_index('Object')['Count']
        num = min(k, len(objects_and_counts))
        return objects_and_counts[:num]
    
    relevant_entities = get_relevant_entities(query)
    
    if len(relevant_entities)<1:
        print("No relevant query results")
        return None
    
    results = query_df(relevant_entities, rel)
    return results

#### Finding all applications of a TiO2

results = query_subrelobj('TiO2', 'CHM-APL')
results

#### Finding all Symmetry Phase Labels of TiO2

results = query_subrelobj('TiO2', 'CHM-SPL')
results

#### Finding all applications of CdTe

results = query_subrelobj('CdTe', 'CHM-APL')
results

#### Finding all Properties of CdTe

results = query_subrelobj('CdTe', 'CHM-PRO')
results

#### Creating a NetworkX Graph

import networkx as nx
import matplotlib.pyplot as plt

# Create an empty directed graph
G = nx.DiGraph()

# Add edges to the graph
for index, row in tqdm(SUBRELOBJ.iterrows(), total = SUBRELOBJ.shape[0]):
    G.add_edge(row['Subject'], row['Object'], relationship=row['Rel'], weight=row['Count'])

# Calculate average degree
average_degree = sum(dict(G.degree()).values()) / float(len(G))
print(f"Average Degree: {average_degree}")

# Corrected calculation of weighted average degree
weighted_degree = sum(attr['weight'] for _, _, attr in G.edges(data=True)) / float(len(G))

print(f"Weighted Average Degree: {weighted_degree}")
