==> 4. SUBRELOBJ.nt Analysis.ipynb <==
import pandas as pd
import random 
from rdflib import Graph, Literal, Namespace, RDF, URIRef, BNode
from tqdm import tqdm
from urllib.parse import quote
from rdflib.plugins.sparql import prepareQuery
import gzip
import pickle

# Load the file SUBRELOBJ.nt.tar.gz from https://zenodo.org/records/10022727

# Initialize RDF graph
G = Graph()

# Open the gzipped .nt file
with gzip.open('SUBRELOBJ.nt.tar.gz', 'rt') as f:
    for line in f:
        try:
            # Load each line into the RDF graph
            G.parse(data=line, format="nt")
        except:
            # Print the bad line for debugging, or just pass to ignore
            pass

terms_dict = pickle.load(open('/home/jupyter/Vineeth/MatKG GitHub pages/RDF Files/entity_uri_mapping.pickle','rb'))

inverse_dict ={value: key for key, value in terms_dict.items()}

# Step 1 & 2: SPARQL query to get rows where subject is 'graphene' and predicate is 'CHM-PRO'

# Step 1: Extract all rows that satisfy the condition (subject = 'graphene' and relationship = 'CHM-PRO')
# Define the SPARQL query
sparql_query = """
PREFIX ex: <http://example.org/>
SELECT ?object ?count WHERE {
    ?index ex:hasSubject <http://example.com/CHM/Graphene> ;
          ex:hasRelationship <http://example.org/CHM-PRO> ;
          ex:hasObject ?object ;
          ex:hasCount ?count .
}
ORDER BY DESC(?count)
"""



# Execute the query
results = g.query(sparql_query)

# Print the results
it =0
for row in results:
    property_, count = row
    print(f"Property: {inverse_dict[property_]}, Count: {count}")
    if it>10:
        break
    it+=1
