==> 6. Linking to Materials Project.ipynb <==
import pandas as pd
import random 
from tqdm import tqdm
import threading
import requests
import time
from pymatgen import MPRester

### Read the CSV file

### Load the file SUBRELOBJ.csv from https://zenodo.org/records/10022727

SUBRELOBJ = pd.read_csv('RDF Files/SUBRELOBJ.csv')

print("Number of rows: {}".format(SUBRELOBJ.shape))
print("Columns : {}".format(SUBRELOBJ.columns.tolist()))

unique_entities = list(set(SUBRELOBJ['Subject'].tolist() + SUBRELOBJ['Object'].tolist()))
print("Number of unique entities : {}".format(len(unique_entities)))

## Query Materials Project

# Define your API key for MPRester
api_key = <your_api_key>

# Function to clean chemical formula strings
def chemical_string_to_formula(chemical_string):
    return chemical_string.replace('(', '').replace(')', '')

# Function to search materials and return material IDs
def search_materials(mpr, chemical_string):
    formula = chemical_string_to_formula(chemical_string)
    try:
        materials = mpr.get_data(formula)
        return [material['material_id'] for material in materials]
    except Exception as e:
        return []

mpr = MPRester(api_key)
mp_id_dict = {}
count =0
for chemical_string in tqdm(unique_entities):
    try:
        material_ids = search_materials(mpr, chemical_string)
        if material_ids:
           # print(f"Material IDs for {chemical_string}: {material_ids}")
            mp_id_dict[chemical_string] = material_ids
        if (count+1)%1000==0:
            print(len(mp_id_dict))
        count+=1
    except:
        pass

len(mp_id_dict)
